+++ src/utils.js
@@ -1,0 +1,13 @@
+function sleep(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+async function rejectAfter(ms, error) {
+  await sleep(ms);
+
+  throw error;
+}
+
+module.exports = {
+  rejectAfter,
+};
